requirements.txt
requests
pycryptodome
sherpa-onnx
numpy
openai
pillow
markdown
